+++ eNMS/controller/inventory.py
@@ -140,10 +140,8 @@ class InventoryController(BaseController):
     def get_session_log(self, session_id):
         return db.fetch("session", id=session_id).content
 
-    def counters(self, property, type):
-        return Counter(
-            str(getattr(instance, property)) for instance in db.fetch_all(type)
-        )
+    def counters(self, property, model):
+        return Counter(v for v, in db.session.query(getattr(models[model], property)))
 
     def export_topology(self, **kwargs):
         workbook = Workbook()